Prepare login for migration
