+++ front/lib/auth.ts
@@ -27,6 +27,10 @@ import { Op } from "sequelize";
 
 import { isDevelopment } from "@app/lib/development";
 import {
+  fetchUserFromSession,
+  maybeUpdateFromExternalUser,
+} from "@app/lib/iam/users";
+import {
   FeatureFlag,
   Key,
   Membership,
@@ -450,88 +454,6 @@ export async function getSession(
   res: NextApiResponse | GetServerSidePropsContext["res"]
 ): Promise<any> {
   return getServerSession(req, res, authOptions);
-}
-
-/**
- * Retrieves the user for a given session
- * @param session any NextAuth session
- * @returns Promise<UserType | null>
- */
-export async function getUserFromSession(
-  session: any
-): Promise<UserTypeWithWorkspaces | null> {
-  if (!session) {
-    return null;
-  }
-
-  const user = await User.findOne({
-    where: {
-      provider: session.provider.provider,
-      providerId: session.provider.id.toString(),
-    },
-  });
-
-  if (!user) {
-    return null;
-  }
-
-  const memberships = await Membership.findAll({
-    where: {
-      userId: user.id,
-      role: { [Op.in]: ["admin", "builder", "user"] },
-    },
-  });
-  const workspaces = await Workspace.findAll({
-    where: {
-      id: memberships.map((m) => m.workspaceId),
-    },
-  });
-
-  if (session.user.image !== user.imageUrl) {
-    void User.update(
-      {
-        imageUrl: session.user.image,
-      },
-      {
-        where: {
-          id: user.id,
-        },
-      }
-    );
-  }
-
-  return {
-    id: user.id,
-    provider: user.provider,
-    username: user.username,
-    email: user.email,
-    firstName: user.firstName,
-    lastName: user.lastName,
-    fullName: user.firstName + (user.lastName ? ` ${user.lastName}` : ""),
-    image: user.imageUrl,
-    workspaces: workspaces.map((w) => {
-      const m = memberships.find((m) => m.workspaceId === w.id);
-      let role = "none" as RoleType;
-      if (m) {
-        switch (m.role) {
-          case "admin":
-          case "builder":
-          case "user":
-            role = m.role;
-            break;
-          default:
-            role = "none";
-        }
-      }
-      return {
-        id: w.id,
-        sId: w.sId,
-        name: w.name,
-        role,
-        segmentation: w.segmentation || null,
-      };
-    }),
-  };
 }
 
 /**
