+++ front/pages/api/user/index.ts
@@ -8,7 +8,8 @@ import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { updateUserFullName } from "@app/lib/api/user";
-import { getSession, getUserFromSession } from "@app/lib/auth";
+import { getSession } from "@app/lib/auth";
+import { getUserFromSession } from "@app/lib/iam/session";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
 export type PostUserMetadataResponseBody = {
