+++ front/pages/index.tsx
@@ -50,7 +50,8 @@ import SimpleSlider from "@app/components/home/carousel";
 import Particles from "@app/components/home/particles";
 import ScrollingHeader from "@app/components/home/scrollingHeader";
 import { PricePlans } from "@app/components/PlansTables";
-import { getSession, getUserFromSession } from "@app/lib/auth";
+import { getSession } from "@app/lib/auth";
+import { getUserFromSession } from "@app/lib/iam/session";
 import { classNames } from "@app/lib/utils";
 import { withGetServerSidePropsLogging } from "@app/logger/withlogging";
